First Draft of Files for the Basic Portfolio
This is a rough draft.  Only the about page has been started but I want to put them in so the base is there for the primary files and then differences can be compared as the files are updated.

index.html contains the html code for the about page of the profile
style.css has all the elements for the about page but only some of the styling is complete for the main section content is complete.